commandOnSelected: CHG: Make "." completely finish selecting (from nested filters), and ".." leave the current filter
This avoids having to enter "." / press Ctrl-D repeatedly to exit from a (potentially deeply nested) filter.
The recursive invocation can clear the filters array to indicate that we're completely done; an added conditional after the recursion breaks out of the loop (via break, not by setting isBreak=t, to avoid the message).
Additional conditions handle the appended "." by clearing filters and then falling through to the case with appended "..".
Adapt the help messages to consider whether filtering is currently in effect.