CHG: Replace optional -f|--file on the template parameter with optional -t|--target on the target parameter
Though it's modeled after "sed -f script-file" and "awk -f program-file" and allows for nice invocation options, the difference is that sed and awk can be found in default locations and therefore can be specified via absolute path in the shebang, whereas for this personal script, the absolute path is very long (and possibly different due to the flexible home directory location). The only way out would be a symlink in /usr/local/bin (error prone and also not 100% stable as an absolute location, as it's just a convention). It's much better to rely on a PATH lookup via /usr/bin/env, but then the command itself becomes the single parsed argument, and the default "-f" argument cannot be specified. (Unless we start relying on new GNU coreutils 8.30 (from 2018, so not even in Ubuntu 16.04)'s new -S option.)
So, let's turn this around, and replace the marking of the template file parameter with an optional marking of the target file. As the embedding of the target file now would take two arguments (-t and TARGET_FILE), also allow the short -tTARGET_FILE and --target=TARGET_FILE parsing variants, if users really want to embed the target file while not using env -S and instead supplying the absolute path. All of that is highly unlikely (but nice to offer, which is why I initially went that route). The benefit of the new solution is that the (easily forgotten) -f is not necessary any longer.