pipelineBuilder: ENH: Allow passing additional commands via --exec [...] ;
As an alternative to -c|--command that doesn't require quoting.
This somewhat eliminates the motivation for --reverse, as any pipeline step can now be specified without additional quoting. However, it can still be useful in an alias or command wrapping use case, where additional arguments are appended (and therefore always apply to the last SIMPLECOMMAND). Rephrase the usage help to suggest that.