ENH: Robustness: Allow partial consumption by default but enable parallel preprocessing via --parallel-preprocess
I've tried to find a way around the SIGPIPE by restructuring the preprocessing pipeline, but that just introduces other (timing) problems.
The problem can be avoided by first capturing the entire input and only then feeding the temp file to the PREPROCESS-COMMAND(s), at the cost of one more cat invocation.
I think it's common enough that preprocessors do not need to consume the entire input, but can stop somewhere after they've parsed enough, and I want this to work out of the box.
On the other hand, this leads to delays when the preprocessing is the slow step, so we have to offer the early parallel processing as well.